requirements.txt
fastapi~=0.115.5
alembic==1.11.1
psycopg2-binary==2.9.10
pydantic==2.7.0
pydantic_settings==2.3.3
python-dotenv==1.0.0
SQLAlchemy==2.0.13
